Move QR code generation functionality to common, as it doesn't rely on Qt at all.
