+++ glue_ar/common/qr.py
